refactor: split SpriteEvents.java into BedSprites.java, DoorScripts.java, InventoryScripts.java, SafeScripts.java and Utilities.java. Moved utility methods to Utilities.java. Minor code refactors to improve readability and other changes not affecting gameplay.
