+++ src/main/java/com/example/escapeyourbedroom/ClickableSprite.java
@@ -16,12 +16,18 @@ class ClickableSprite extends ImageView {
         // Set the displayed image
         normalImage = new Image(imagePath);
         setImage(normalImage);
+
         this.name = name;
         x = spriteX;
         y = spriteY;
+
         // setTranslate moves the image from [0,0], which is the middle of the screen. Apparently it's the only way to move the image because the other two don't work? Classic javaFX shenanigans
         setTranslateX(x);
         setTranslateY(y);
+
+        // Automatically try to assign zoomed image
+        zoomedImage = new Image(Utilities.zoomifyName(imagePath));
+
         root.getChildren().add(this);
     }
     void hide() {
@@ -32,31 +38,27 @@ class ClickableSprite extends ImageView {
         toFront();
     }
     void setHighlightOnHover() {
-
-        // For something funny, try to put your mouse on the edge of a sprite (try different edges, only works on some)
-
-        // TODO: Make highlighting add a white outline to the sprite instead of resizing it (might be difficult)
         setOnMouseEntered(event -> {
             if (!isZoomed) {
                 // When mouse enters a sprite, show its name on the name tag and highlight it
-                nameTag.setText(name);
+                Utilities.nameTag.setText(name);
                 setScaleX(this.getScaleX() * 1.1);
                 setScaleY(this.getScaleY() * 1.1);
-                nameTag.show();
+                Utilities.nameTag.show();
             }
         });
         setOnMouseExited(event -> {
             if (!isZoomed) {
                 // When the mouse exits the sprite, hide the name tag and stop highlighting it
-                nameTag.hide();
+                Utilities.nameTag.hide();
                 setScaleX(1);
                 setScaleY(1);
             }
         });
         setOnMouseMoved(mouseEvent -> {
             // If the sprite its zoomed in, make the name tag follow the cursor; otherwise hide it
-            if (!isZoomed) nameTag.setPosToCursor(mouseEvent.getSceneX(),  mouseEvent.getSceneY());
-            else nameTag.hide();
+            if (!isZoomed) Utilities.nameTag.setPosToCursor(mouseEvent.getSceneX(),  mouseEvent.getSceneY());
+            else Utilities.nameTag.hide();
         });
     }
 
@@ -65,20 +67,20 @@ class ClickableSprite extends ImageView {
         setOnMouseEntered(event -> {
             if (!isZoomed) {
                 // When mouse enters a sprite, show its name on the name tag and highlight it
-                nameTag.setText(name);
-                nameTag.show();
+                Utilities.nameTag.setText(name);
+                Utilities.nameTag.show();
             }
         });
         setOnMouseExited(event -> {
             if (!isZoomed) {
                 // When the mouse exits the sprite, hide the name tag and stop highlighting it
-                nameTag.hide();
+                Utilities.nameTag.hide();
             }
         });
         setOnMouseMoved(mouseEvent -> {
             // If the sprite its zoomed in, make the name tag follow the cursor; otherwise hide it
-            if (!isZoomed) nameTag.setPosToCursor(mouseEvent.getSceneX(),  mouseEvent.getSceneY());
-            else nameTag.hide();
+            if (!isZoomed) Utilities.nameTag.setPosToCursor(mouseEvent.getSceneX(),  mouseEvent.getSceneY());
+            else Utilities.nameTag.hide();
         });
     }
 
@@ -101,17 +103,18 @@ class ClickableSprite extends ImageView {
     }
 
     void setParentScene(int sceneIndex) {
+        // Add sprite to its appropriate list and hide it if it shouldn't be displayed
         sceneSprites.get(sceneIndex).add(this);
         if (sceneIndex != currentScene) hide();
     }
 
-    // The name needs some refinement
     void zoomInto() {
         // Revert scale back to the un-zoomed one so that the sprite isn't growing when zooming and un-zooming
         setScaleX(1);
         setScaleY(1);
+
         // Hide all sprites on the current scene
-        updateSpritesVisibility(currentScene, false);
+        Utilities.updateSpritesVisibility(currentScene, false);
 
         // Renders the zoomed-in item
         setImage(zoomedImage);
@@ -122,16 +125,16 @@ class ClickableSprite extends ImageView {
     }
     void zoomOut() {
         // Show all sprites on the current scene
-        updateSpritesVisibility(currentScene, true);
+        Utilities.updateSpritesVisibility(currentScene, true);
         setImage(this.normalImage);
         setTranslateX(x);
         setTranslateY(y);
-        exitButton.hide(); // Hide the exit button on zoom-out
+
+        // Hide the exit button on zoom-out
+        Utilities.exitButton.hide();
+
         if (childSprite != null) childSprite.hide();
         isZoomed = false;
 
-    }
-    void setZoomedImage(String imagePath) {
-        zoomedImage = new Image(imagePath);
     }
 }
